refactor: extract hook (review comment)
